Completly migrated from MVVM to code behind
